handler: _update_step: remove calling of `mark_selected`
This should not be necessary in general, since it gets triggered via
`CursorMoved` already.
It also might trigger an error due to `self._vim.current.window` not
being a valid window anymore, when `cursor` gets accessed [1].
It also should not be using the global current window's cursor (which
might be for another buffer!) to mark selected nodes in the buffer.

1: https://github.com/numirias/semshi/issues/72